feat(SiteHeader): split the header into separate component for better maintainability
